SEASON 3 OUT NOW!
 # SEASON 3 OUT NOW!

 ## Summary
A long time in the making, we're proud to release season three! For the sake of getting it out early, it's unfortunately unfinished.
But this was an ambitious update, full of optimizations and roadblocks! We feel that you will still enjoy this update and be excited what's in store
This is a revamp of the current season 2, consisting of a new
 - Search bar
 - Game page
 - File formatting
 - Game page compiling
 - Optimized files for faster load time
 - More detailed info on game pages
 - Cleaner code
 - etc.

 ## Known issues
 - Some games are currently broken (45, thanks rm) and do not have proper iframe configurations. Should be fixed by friday
 - Descriptions are unfinished
 - Python is clunky and hard to understand